+++ apps/commerce-api/src/main/java/com/loopers/application/point/PointCriteria.java
@@ -2,11 +2,13 @@ package com.loopers.application.point;
 
 import com.loopers.domain.point.PointCommand;
 
+import java.math.BigDecimal;
+
 public class PointCriteria {
 
     public record Charge(
             String loginId,
-            Long amount
+            BigDecimal amount
     ) {
         public PointCommand.Charge toCommand() {
             return new PointCommand.Charge(loginId, amount);
