+++ apps/commerce-api/src/main/java/com/loopers/domain/point/PointInfo.java
@@ -2,10 +2,12 @@ package com.loopers.domain.point;
 
 import com.loopers.domain.user.LoginId;
 
+import java.math.BigDecimal;
+
 public record PointInfo(
         Long id,
         LoginId loginId,
-        Long amount
+        BigDecimal amount
 ) {
     public static PointInfo from(final PointEntity pointEntity) {
         return new PointInfo(
